pyproject.toml
[project]
name = "learn-ibis"
version = "0.1.0"
description = "Snippets and examples to improve my understanding of ibis and databases"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ibis-framework[duckdb,postgres]>=10.6.0",
    "ipykernel>=6.29.5",
    "pandas>=2.3.0",
    "psycopg2-binary>=2.9.10",
    "pyarrow>=20.0.0",
    "pydantic-settings>=2.0.0",
]

[project.optional-dependencies]
dev = ["ruff>=0.12.0", "ty>=0.0.1a11"]
test = ["pytest>=8.4.1"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build]
packages = ["src/learn_ibis"]
